drivers/mtd/bch: fix size_t overflow when offset > 4GB

+++ drivers/bch/bchlib_write.c
@@ -50,7 +50,7 @@
  *
  ****************************************************************************/
 
-ssize_t bchlib_write(FAR void *handle, FAR const char *buffer, size_t offset,
+ssize_t bchlib_write(FAR void *handle, FAR const char *buffer, off_t offset,
         size_t len)
 {
   FAR struct bchlib_s *bch = (FAR struct bchlib_s *)handle;
